requirements.txt
Django==1.8.9
django-jsonfield==0.9.19

psycopg2==2.6.1
pyvmomi
boto3
django-crontab
python-dateutil
redis
qingcloud-sdk
mysqlclient==1.3.9
Jinja2
django-extensions
django-log-request-id
sqlalchemy==1.1.10
kubernetes==2.0.0
schema==0.6.6

pycrypto
google-api-python-client==1.4.1
djangorestframework==3.3.2
markdown=2.6.11
django-filter=1.1.0
django-storages=1.6.6